pyproject.toml
[project]
name = "uwf-rag-experiments"
version = "0.1.0"
description = "Reproducible RAG experiments for the UWF advising chatbot: baseline single-agent pipeline + RAGAS evaluation + retrieval ablations."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "LoganSelner", email = "loganselner@gmail.com" }
]
requires-python = ">=3.11,<3.12"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-chroma>=1.1.0",
    "langchain-text-splitters>=1.1.0",

    "pypdf>=6.7.0",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "typer>=0.23.1",
]

[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "pytest",
    "pre-commit",
    "types-pyyaml>=6.0.12.20250915",
]

[project.scripts]
rag = "rag_experiments.cli:app"

[build-system]
requires = ["hatchling>=1.25"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/rag_experiments"]

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP", "B", "RUF"]
ignore = ["B008"]

[tool.ruff.lint.isort]
known-first-party = ["rag_experiments"]
combine-as-imports = true
force-sort-within-sections = true

[tool.mypy]
python_version = "3.11"
warn_unused_ignores = true
warn_redundant_casts = true
warn_unreachable = true
disallow_untyped_defs = true
no_implicit_optional = true
strict_optional = true
exclude = '^tests/'
mypy_path = ["src"]
files = ["src"]
explicit_package_bases = true

[tool.pytest.ini_options]
addopts = "-q -ra --strict-markers"
testpaths = ["tests"]
